refactor to separate logic from ceilometer-dependent components + base for sourcedict processing testing
